% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{col_character}
\alias{col_character}
\title{Colour text in an character-based theme.}
\usage{
col_character(char, text = NULL)
}
\arguments{
\item{char}{The character colour to use.}

\item{text}{The text to theme (or \code{NULL} if you just want to use "char").}
}
\value{
A defect-themed colour function.
}
\description{
Colour text in an character-based theme.
}
\examples{
col_character("Ironclad")
col_character("Silent", "Other text")
}
