% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{col_defect}
\alias{col_defect}
\title{Colour text in an Defect-based theme.}
\usage{
col_defect(...)
}
\arguments{
\item{...}{Character strings, they will be pasted together with paste0(), before applying the style function.}
}
\value{
A defect-themed colour function.
}
\description{
Colour text in an Defect-based theme.
}
\examples{
col_defect("Defect-themed text")
}
