% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{col_regent}
\alias{col_regent}
\title{Colour text in an Regent-based theme.}
\usage{
col_regent(...)
}
\arguments{
\item{...}{Character strings, they will be pasted together with paste0(), before applying the style function.}
}
\value{
A regent-themed colour function.
}
\description{
Colour text in an Regent-based theme.
}
\examples{
col_regent("Regent-themed text")
}
