% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{col_silent}
\alias{col_silent}
\title{Colour text in an Silent-based theme.}
\usage{
col_silent(...)
}
\value{
A silent-themed colour function.
}
\description{
Colour text in an Silent-based theme.
}
\examples{
col_silent("Silent-themed text")
}
