% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsts-package.R
\name{.ignore_unused_imports}
\alias{.ignore_unused_imports}
\title{Force R CMD CHECK to treat packages as imported.}
\usage{
.ignore_unused_imports()
}
\value{
Nothing
}
\description{
Force R CMD CHECK to treat packages as imported.
}
\note{
This allows us to force dependencies to be considered in the package (particularly necessary with R6 classes).
}
\keyword{internal}
