% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsts-utils.R
\name{format_sts2id}
\alias{format_sts2id}
\title{Format sts2 ids to a prettier form.}
\usage{
format_sts2id(id)
}
\arguments{
\item{id}{}
}
\value{
The formatted ids.
}
\description{
Format sts2 ids to a prettier form.
}
\note{
This assumes that ids have only 1 "." character.
}
\keyword{internal}
