% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsts-utils.R
\name{get_field}
\alias{get_field}
\title{Get field from a list of runs (or players)}
\usage{
get_field(x, fieldname)
}
\arguments{
\item{x}{A list of \link{STS2Run} or \link{STS2Player} objects.}

\item{fieldname}{The name of the field to retrieve.}
}
\value{
The formatted ids.
}
\description{
Get field from a list of runs (or players)
}
\keyword{internal}
