% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sts_history.R
\name{load_sts_history}
\alias{load_sts_history}
\title{Load Slay the Spire 2 Run Data}
\usage{
load_sts_history(path = NULL, profilenum = NULL, game = 2)
}
\arguments{
\item{path}{The path to your installation.}

\item{game}{Which game to retrieve run data for (2 = STS2, 1 = STS), currently only STS2 is implemented.}

\item{profile_num}{Which profile to retrieve data for (or NULL to retrieve all).}
}
\value{
The parsed run data.
}
\description{
Load Slay the Spire 2 Run Data
}
\examples{
load_sts2_history()
}
