% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsts-package.R
\docType{package}
\name{rsts-package}
\alias{rsts}
\alias{rsts-package}
\title{rsts: R Slay The Spire Interface}
\description{
R6 classes and helpful functions for working with Slay The Spire run histories.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fwimp/rsts}
  \item \url{https://fwimp.github.io/rsts/}
  \item Report bugs at \url{https://github.com/fwimp/rsts/issues}
}

}
\author{
\strong{Maintainer}: Francis Windram \email{francis.windram17@imperial.ac.uk} [copyright holder]

}
\keyword{internal}
