% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{STS2Deck}
\alias{STS2Deck}
\title{R6 Class representing a set of Slay the Spire 2 Cards.}
\value{
A new \code{STS2Cards} object.
}
\description{
This is the general holding class for a set of sts2 Cards.
It stores the main card data for the player.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{cards}}{A list of cards.}

\item{\code{floorfound}}{The floors upon which the cards were found.}

\item{\code{upgradelevel}}{The upgrade level of each card.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-STS2Deck-new}{\code{STS2Deck$new()}}
\item \href{#method-STS2Deck-clone}{\code{STS2Deck$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STS2Deck-new"></a>}}
\if{latex}{\out{\hypertarget{method-STS2Deck-new}{}}}
\subsection{Method \code{new()}}{
Create a new relics object from player data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STS2Deck$new(carddata)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{carddata}}{The subset of the list output from jsonlite, usually passed in via \code{STS2Player}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STS2Deck-clone"></a>}}
\if{latex}{\out{\hypertarget{method-STS2Deck-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STS2Deck$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
