% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sts2map.R
\name{STS2Floor}
\alias{STS2Floor}
\title{Slay the Spire 2 floor (R6).}
\value{
A new \link{STS2Floor} object.
}
\description{
This is the general holding class for a sts2 floor.
}
\concept{R6}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{act}}{The act this floor appeared in.}

\item{\code{map}}{The \link{STS2Map} object this floor belongs to.}

\item{\code{floor_type}}{The type of floor that this is.}

\item{\code{player_stats}}{The stats of the player/s at the end of this floor. A list of \link{STS2PlayerMidrun} objects.}

\item{\code{turns_taken}}{The number of turns this room took.}

\item{\code{model_id}}{The model id of the encounter (or \code{NULL} if rest site/treasure room).}

\item{\code{monsters}}{The monsters present on the floor.}

\item{\code{rooms}}{A list of the unparsed room data.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-STS2Floor-new}{\code{STS2Floor$new()}}
\item \href{#method-STS2Floor-clone}{\code{STS2Floor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STS2Floor-new"></a>}}
\if{latex}{\out{\hypertarget{method-STS2Floor-new}{}}}
\subsection{Method \code{new()}}{
Create a new floor object from floor data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STS2Floor$new(floordata, act = 0, map = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{floordata}}{The individual floor-level output from jsonlite, usually passed in via \link{STS2Map}.}

\item{\code{act}}{The act this floor appeared in.}

\item{\code{map}}{The \link{STS2Map} object this floor belongs to.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STS2Floor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-STS2Floor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STS2Floor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
