% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sts2map.R
\name{STS2Map}
\alias{STS2Map}
\title{Map points in a Slay The Spire 2 run (R6).}
\value{
A new \code{STS2Map} object.
}
\description{
This is the general holding class for STS2 map points.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{run}}{The run}

\item{\code{floors}}{The floors in the map. A list of \code{STS2Floor} objects.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-STS2Map-new}{\code{STS2Map$new()}}
\item \href{#method-STS2Map-clone}{\code{STS2Map$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STS2Map-new"></a>}}
\if{latex}{\out{\hypertarget{method-STS2Map-new}{}}}
\subsection{Method \code{new()}}{
Create a new map object from map data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STS2Map$new(mapdata, run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mapdata}}{The subset of the list output from jsonlite, usually passed in via \code{STS2Player}.}

\item{\code{run}}{The \code{STS2Run} object this map belongs to.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STS2Map-clone"></a>}}
\if{latex}{\out{\hypertarget{method-STS2Map-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STS2Map$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
