% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sts2_classes.R
\name{STS2Player}
\alias{STS2Player}
\title{Slay the Spire 2 player (R6).}
\value{
A new \code{STS2Player} object.
}
\description{
This is the general holding class for an sts2 player.
It stores the main player data for the run.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{run}}{The \code{STS2Run} object that this player object originates from.}

\item{\code{playercharacter}}{The character that the player was playing}

\item{\code{deck}}{The deck of the player at the end of the run. An \code{STS2Deck} object.}

\item{\code{id}}{The steam player id or 1 if single player.}

\item{\code{max_potion_slot_count}}{The maximum number of potions a player could hold.}

\item{\code{potions}}{The potions held by the player at the end of the run.}

\item{\code{relics}}{The relics held by the player at the end of the run. An \code{STS2Relics} object}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{max_health}}{The max health of the character at the end of the run.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-STS2Player-new}{\code{STS2Player$new()}}
\item \href{#method-STS2Player-clone}{\code{STS2Player$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STS2Player-new"></a>}}
\if{latex}{\out{\hypertarget{method-STS2Player-new}{}}}
\subsection{Method \code{new()}}{
Create a new run object from player data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STS2Player$new(playerdata, run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{playerdata}}{The subset of the list output from jsonlite, usually passed in via \code{STS2Run}.}

\item{\code{run}}{The STS2Run object that this player object originates from.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STS2Player-clone"></a>}}
\if{latex}{\out{\hypertarget{method-STS2Player-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STS2Player$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
