% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sts2map.R
\name{STS2PlayerMidrun}
\alias{STS2PlayerMidrun}
\title{A Slay the Spire 2 player, mid-run (R6).}
\value{
A new \link{STS2PlayerMidrun} object.
}
\description{
This is the general holding class for a sts2 player at a point in the run.
}
\note{
All fields in this are considered to be "at the end of this floor".
For example current_hp on a floor where a player dies will be 0, even if they entered the fight with more than 0 HP.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{floor}}{The \link{STS2Floor} object this midrun object refers to.}

\item{\code{player}}{The \link{STS2Player} object this midrun object refers to.}

\item{\code{current_gold}}{Current gold.}

\item{\code{current_hp}}{Current health.}

\item{\code{damage_taken}}{Damage taken during the floor.}

\item{\code{gold_gained}}{Gold gained.}

\item{\code{gold_lost}}{Gold lost.}

\item{\code{gold_spent}}{Gold spent.}

\item{\code{gold_stolen}}{Gold stolen (damn gremlins).}

\item{\code{hp_healed}}{Health recovered during the floor.}

\item{\code{max_hp}}{Max health.}

\item{\code{max_hp_gained}}{Max health gained during the floor.}

\item{\code{max_hp_lost}}{Max health lost during the floor.}

\item{\code{player_id}}{The id of the player within the run data.}

\item{\code{ancient_choice}}{The choices available from the ancient (and which was picked).}

\item{\code{bought_colorless}}{The colorless cards bought during the floor.}

\item{\code{bought_potions}}{The potions bought during the floor.}

\item{\code{bought_relics}}{The relics bought during the floor.}

\item{\code{card_choices}}{The card choices available.}

\item{\code{cards_enchanted}}{The cards enchanted during the floor.}

\item{\code{cards_gained}}{The cards gained from the floor.}

\item{\code{cards_removed}}{The cards removed during the floor.}

\item{\code{cards_transformed}}{The cards transformed during the floor.}

\item{\code{completed_quests}}{The quests completed on the floor.}

\item{\code{event_choices}}{The choices made during the event (not in a particularly useful form).}

\item{\code{potion_choices}}{The potion choices available.}

\item{\code{potion_discarded}}{The potions discarded during the floor.}

\item{\code{potion_used}}{The potions used during the floor.}

\item{\code{relic_choices}}{The choices available from the floor (and which was picked).}

\item{\code{relics_removed}}{The relics removed during the floor.}

\item{\code{rest_site_choices}}{The rest site choices (and which was picked).}

\item{\code{upgraded_cards}}{The cards upgraded during the floor.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-STS2PlayerMidrun-new}{\code{STS2PlayerMidrun$new()}}
\item \href{#method-STS2PlayerMidrun-clone}{\code{STS2PlayerMidrun$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STS2PlayerMidrun-new"></a>}}
\if{latex}{\out{\hypertarget{method-STS2PlayerMidrun-new}{}}}
\subsection{Method \code{new()}}{
Create a new mid-run player object from player data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STS2PlayerMidrun$new(playerstats, floor = NULL, player = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{playerstats}}{The individual player_stats output from jsonlite, usually passed in via \link{STS2Player}.}

\item{\code{floor}}{The \link{STS2Floor} object this midrun object refers to.}

\item{\code{player}}{The \link{STS2Player} object this midrun object refers to.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STS2PlayerMidrun-clone"></a>}}
\if{latex}{\out{\hypertarget{method-STS2PlayerMidrun-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STS2PlayerMidrun$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
