% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sts2_classes.R
\name{STS2Run}
\alias{STS2Run}
\title{Slay the Spire 2 run (R6).}
\value{
A new \code{STS2Run} object.

Nothing (called for side-effect)

An \code{STS2Player} object or \code{NULL}.

A list of \code{STS2Player} objects containing only selected characters.

The outcome of the run as a string (one of Win, Abandoned, or Loss).
}
\description{
This is the general holding class for an sts2 run.
It stores the main metadata and lists containing more in-depth run data.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{acts}}{A vector of the acts that the player encountered.}

\item{\code{ascension}}{The ascension level of the run.}

\item{\code{build_id}}{The build of the game that this run was generated from.}

\item{\code{game_mode}}{The game mode that the player was playing.}

\item{\code{killed_by_encounter}}{The encounter that killed the player.}

\item{\code{killed_by_event}}{The event that killed the player.}

\item{\code{map}}{A list storing the play-by-play of the run. An \code{STS2Map} object.}

\item{\code{modifiers}}{Any modifiers that the player had turned on.}

\item{\code{platform_type}}{The platform of the game (e.g. steam).}

\item{\code{players}}{A list of player information. A list of \code{STS2Player} objects.}

\item{\code{run_time}}{The run duration (in seconds).}

\item{\code{schema_version}}{The run schema version.}

\item{\code{seed}}{The seed of the run.}

\item{\code{start_time}}{The start time (in seconds from unix epoch, UTC).}

\item{\code{was_abandoned}}{Whether the player abandoned the run.}

\item{\code{win}}{Whether the player won the run.}

\item{\code{ownerid}}{The steam ID of the run owner.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-STS2Run-new}{\code{STS2Run$new()}}
\item \href{#method-STS2Run-print}{\code{STS2Run$print()}}
\item \href{#method-STS2Run-get_individual_player_data}{\code{STS2Run$get_individual_player_data()}}
\item \href{#method-STS2Run-get_character}{\code{STS2Run$get_character()}}
\item \href{#method-STS2Run-get_outcome}{\code{STS2Run$get_outcome()}}
\item \href{#method-STS2Run-clone}{\code{STS2Run$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STS2Run-new"></a>}}
\if{latex}{\out{\hypertarget{method-STS2Run-new}{}}}
\subsection{Method \code{new()}}{
Create a new run object from data parsed with jsonlite.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STS2Run$new(rundata, steamid = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rundata}}{The list output from jsonlite containing the run data.}

\item{\code{steamid}}{The owner's steamid (for differentiating in the case of multiplayer runs).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STS2Run-print"></a>}}
\if{latex}{\out{\hypertarget{method-STS2Run-print}{}}}
\subsection{Method \code{print()}}{
Print an \code{STS2Run} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STS2Run$print(..., full = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Unused.}

\item{\code{full}}{Whether to print extra internal run information.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STS2Run-get_individual_player_data"></a>}}
\if{latex}{\out{\hypertarget{method-STS2Run-get_individual_player_data}{}}}
\subsection{Method \code{get_individual_player_data()}}{
Retrieve player data for a given player from a run.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STS2Run$get_individual_player_data(id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the Steam ID of the player data to retrieve (or \code{NULL} to retrieve the data of the run owner).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STS2Run-get_character"></a>}}
\if{latex}{\out{\hypertarget{method-STS2Run-get_character}{}}}
\subsection{Method \code{get_character()}}{
Retrieve data for a character.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STS2Run$get_character(char)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{char}}{The character/s to retrieve data for.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STS2Run-get_outcome"></a>}}
\if{latex}{\out{\hypertarget{method-STS2Run-get_outcome}{}}}
\subsection{Method \code{get_outcome()}}{
Get run outcome.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STS2Run$get_outcome()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STS2Run-clone"></a>}}
\if{latex}{\out{\hypertarget{method-STS2Run-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STS2Run$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
