% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topcards.R
\name{topcards}
\alias{topcards}
\title{Retrieve the most-taken cards across a history of runs.}
\usage{
topcards(
  runhistory,
  n = 10,
  char = NULL,
  return_unique = FALSE,
  ignore_basics = TRUE
)
}
\arguments{
\item{runhistory}{An \link{STS2RunHistory} object, filtered if necessary.}

\item{n}{The number of results to retrieve, or \code{NULL} to retrieve all.}

\item{char}{The character/s to filter by (or \code{NULL} to return all).}

\item{return_unique}{Return only 1 entry per unique card in a run. This gives more of a metric of "runs where this card was in your deck".}

\item{ignore_basics}{Don't consider basic cards (strike/defend).}
}
\value{
A \code{data.frame} of cards and the number of times they appeared in a run
}
\description{
Retrieve the most-taken cards across a history of runs.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
myruns <- load_sts_history()
topcards(myruns)
topcards(myruns$filter_outcome("win"), n = 20, return_unique = TRUE)
\dontshow{\}) # examplesIf}
}
