% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winrate.R
\name{winrate}
\alias{winrate}
\title{Calculate (and potentially plot) winrates by character.}
\usage{
winrate(
  runhistory,
  plotit = TRUE,
  relative = FALSE,
  expected = NULL,
  ci = TRUE,
  samples = 200,
  lower_quant = 0.025,
  upper_quant = 0.975
)
}
\arguments{
\item{runhistory}{An \link{STS2RunHistory} object, filtered if necessary.}

\item{plotit}{Whether to plot the data or just return a summary \code{data.frame}}

\item{relative}{Whether to calculate the character-based winrate relative to the total winrate.}

\item{expected}{If set, the expected winrate to scale values relative to (automatically turns on \code{relative} when set.)}

\item{ci}{Whether to calculate and plot 95\% confidence intervals}

\item{samples}{The number of iterations to use when generating CIs.}

\item{lower_quant}{The lower quantile of the CIs.}

\item{upper_quant}{The upper quantile of the CIs.}
}
\value{
A \code{data.frame} summary of winrates by character.
}
\description{
Calculate (and potentially plot) winrates by character.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
myruns <- load_sts_history()
winrate(myruns)
\dontshow{\}) # examplesIf}
}
