% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{col_ironclad}
\alias{col_ironclad}
\title{Colour text in an Ironclad-based theme.}
\usage{
col_ironclad(...)
}
\arguments{
\item{...}{Character strings, they will be pasted together with paste0(), before applying the style function.}
}
\value{
An ironclad-themed colour function.
}
\description{
Colour text in an Ironclad-based theme.
}
\examples{
col_ironclad("Ironclad-themed text")
}
