% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{col_necrobinder}
\alias{col_necrobinder}
\title{Colour text in an Necrobinder-based theme.}
\usage{
col_necrobinder(...)
}
\value{
A necrobinder-themed colour function.
}
\description{
Colour text in an Necrobinder-based theme.
}
\examples{
col_necrobinder("Necrobinder-themed text")
}
