% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sts_history.R
\name{load_sts_history}
\alias{load_sts_history}
\title{Load Slay the Spire 2 Run Data}
\usage{
load_sts_history(
  path = NULL,
  id = NULL,
  profilenum = 1,
  game = 2,
  platform = c("windows", "mac", "linux"),
  players = NULL
)
}
\arguments{
\item{path}{The path to your installation.}

\item{id}{The Steam ID of the data you want to retrieve. If \code{NULL}, will just get the first ID directory when \code{path = NULL}.}

\item{profilenum}{Which profile to retrieve data for (or \code{NULL} to retrieve all).}

\item{game}{Which game to retrieve run data for (2 = STS2, 1 = STS), currently only STS2 is implemented.}

\item{platform}{Which platform are you running on?}

\item{players}{Only analyse runs with this number of players (or \code{NULL} for all, default).}
}
\value{
The parsed run data.
}
\description{
Load Slay the Spire 2 Run Data
}
\note{
If you provide a range of numbers for the \code{players} argument e.g. \code{1:3}, any number of players in this range will be included.

If you provide a path yourself, make sure to provide a steam ID to consider as the "owning" player. This is necessary for any functions that automatically get player data from multiplayer runs.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
load_sts_history()
\dontshow{\}) # examplesIf}
}
