% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sts2_classes.R
\name{STS2Relics}
\alias{STS2Relics}
\title{Set of Slay the Spire 2 relics (R6).}
\value{
A new \code{STS2Relics} object.

Nothing (called for side-effect)
}
\description{
This is the general holding class for a set of sts2 relics.
It stores the main relic data for the player.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{player}}{The \code{STS2Player} object these relics belong to.}

\item{\code{relicnames}}{A list of relics.}

\item{\code{floorfound}}{The floors upon which the relics were found.}

\item{\code{extra_data}}{The extra_data associated with the relics (unparsed)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-STS2Relics-new}{\code{STS2Relics$new()}}
\item \href{#method-STS2Relics-print}{\code{STS2Relics$print()}}
\item \href{#method-STS2Relics-clone}{\code{STS2Relics$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STS2Relics-new"></a>}}
\if{latex}{\out{\hypertarget{method-STS2Relics-new}{}}}
\subsection{Method \code{new()}}{
Create a new relics object from player data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STS2Relics$new(relicdata, player = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{relicdata}}{The subset of the list output from jsonlite, usually passed in via \code{STS2Player}.}

\item{\code{player}}{The \code{STS2Player} object these relics belong to.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STS2Relics-print"></a>}}
\if{latex}{\out{\hypertarget{method-STS2Relics-print}{}}}
\subsection{Method \code{print()}}{
Print an \code{STS2Relics} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STS2Relics$print(..., floor = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Unused.}

\item{\code{floor}}{Whether to show the floor on which a card was obtained.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STS2Relics-clone"></a>}}
\if{latex}{\out{\hypertarget{method-STS2Relics-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STS2Relics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
