% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sts2_classes.R
\name{STS2RunHistory}
\alias{STS2RunHistory}
\title{Slay the Spire 2 run history (R6).}
\value{
A new \code{STS2RunHistory} object.

Nothing (called for side-effect)
}
\description{
This is the general holding class for a full history of sts2 runs.

It can be indexed like a list (i.e. \code{x[1]} or \code{x[[1]]}).
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{runs}}{The parsed run logs. A list of \code{STS2Run} objects.}

\item{\code{ownerid}}{The steam ID of the run history owner.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-STS2RunHistory-new}{\code{STS2RunHistory$new()}}
\item \href{#method-STS2RunHistory-print}{\code{STS2RunHistory$print()}}
\item \href{#method-STS2RunHistory-get_individual_player_data}{\code{STS2RunHistory$get_individual_player_data()}}
\item \href{#method-STS2RunHistory-clone}{\code{STS2RunHistory$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STS2RunHistory-new"></a>}}
\if{latex}{\out{\hypertarget{method-STS2RunHistory-new}{}}}
\subsection{Method \code{new()}}{
Create a new run history container from a list of \code{STS2Run} objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STS2RunHistory$new(historydata, steamid = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{historydata}}{A list of \code{STS2Run} objects.}

\item{\code{steamid}}{The owner's steamid (for differentiating in the case of multiplayer runs).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STS2RunHistory-print"></a>}}
\if{latex}{\out{\hypertarget{method-STS2RunHistory-print}{}}}
\subsection{Method \code{print()}}{
Print an \code{STS2RunHistory} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STS2RunHistory$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Unused.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STS2RunHistory-get_individual_player_data"></a>}}
\if{latex}{\out{\hypertarget{method-STS2RunHistory-get_individual_player_data}{}}}
\subsection{Method \code{get_individual_player_data()}}{
Retrieve player data for a given player from runs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STS2RunHistory$get_individual_player_data(id = NULL, excludemissing = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the Steam ID of the player data to retrieve (or \code{NULL} to retrieve the data of the run owner).}

\item{\code{excludemissing}}{Exclude entries from the list where the desired player is not present. (This will result in a list that may be shorter than the number of runs in the history.)

@returns A list of \code{STS2Player} objects.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STS2RunHistory-clone"></a>}}
\if{latex}{\out{\hypertarget{method-STS2RunHistory-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STS2RunHistory$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
