% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sts2runhistory.R
\name{STS2RunHistory}
\alias{STS2RunHistory}
\title{Slay the Spire 2 run history (R6).}
\value{
A new \link{STS2RunHistory} object.
}
\description{
This is the general holding class for a full history of sts2 runs.

It can be indexed like a list (i.e. \code{x[1]} or \code{x[[1]]}).
}
\section{Filters}{

Any method starting \code{filter_} returns a new \link{STS2RunHistory} object.

As such these can be chained together:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{myruns <- load_sts_history()
myruns$filter_ascension(0)$filter_outcome("Win")
}\if{html}{\out{</div>}}
}

\section{STS2Run objects}{

\link{STS2Run} objects are passed by reference. As such if you modify a run in a filtered history, those changes will appear in the original list.
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{runs}}{The parsed run logs. A list of \link{STS2Run} objects.}

\item{\code{ownerid}}{The steam ID of the run history owner.}

\item{\code{filtersteps}}{The filtering steps performed on this run history.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-STS2RunHistory-new}{\code{STS2RunHistory$new()}}
\item \href{#method-STS2RunHistory-print}{\code{STS2RunHistory$print()}}
\item \href{#method-STS2RunHistory-get_individual_player_data}{\code{STS2RunHistory$get_individual_player_data()}}
\item \href{#method-STS2RunHistory-get_player_data}{\code{STS2RunHistory$get_player_data()}}
\item \href{#method-STS2RunHistory-filter_seed}{\code{STS2RunHistory$filter_seed()}}
\item \href{#method-STS2RunHistory-filter_character}{\code{STS2RunHistory$filter_character()}}
\item \href{#method-STS2RunHistory-filter_outcome}{\code{STS2RunHistory$filter_outcome()}}
\item \href{#method-STS2RunHistory-filter_ascension}{\code{STS2RunHistory$filter_ascension()}}
\item \href{#method-STS2RunHistory-filter_playercount}{\code{STS2RunHistory$filter_playercount()}}
\item \href{#method-STS2RunHistory-filter_version}{\code{STS2RunHistory$filter_version()}}
\item \href{#method-STS2RunHistory-filter_floorcount}{\code{STS2RunHistory$filter_floorcount()}}
\item \href{#method-STS2RunHistory-generate_summary}{\code{STS2RunHistory$generate_summary()}}
\item \href{#method-STS2RunHistory-clone}{\code{STS2RunHistory$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STS2RunHistory-new"></a>}}
\if{latex}{\out{\hypertarget{method-STS2RunHistory-new}{}}}
\subsection{Method \code{new()}}{
Create a new run history container from a list of \link{STS2Run} objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STS2RunHistory$new(historydata, steamid = NULL, filtersteps = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{historydata}}{A list of \link{STS2Run} objects.}

\item{\code{steamid}}{The owner's steamid (for differentiating in the case of multiplayer runs).}

\item{\code{filtersteps}}{The filtering steps performed on this run history.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STS2RunHistory-print"></a>}}
\if{latex}{\out{\hypertarget{method-STS2RunHistory-print}{}}}
\subsection{Method \code{print()}}{
Print an \link{STS2RunHistory} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STS2RunHistory$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Unused.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STS2RunHistory-get_individual_player_data"></a>}}
\if{latex}{\out{\hypertarget{method-STS2RunHistory-get_individual_player_data}{}}}
\subsection{Method \code{get_individual_player_data()}}{
Retrieve player data for a given player from runs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STS2RunHistory$get_individual_player_data(id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The Steam ID of the player data to retrieve (or \code{NULL} to retrieve the data of the run owner).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STS2RunHistory-get_player_data"></a>}}
\if{latex}{\out{\hypertarget{method-STS2RunHistory-get_player_data}{}}}
\subsection{Method \code{get_player_data()}}{
Retrieve player data for all players from runs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STS2RunHistory$get_player_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STS2RunHistory-filter_seed"></a>}}
\if{latex}{\out{\hypertarget{method-STS2RunHistory-filter_seed}{}}}
\subsection{Method \code{filter_seed()}}{
Retrieve runs by seed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STS2RunHistory$filter_seed(seed, .filtertext = "filtered by seed")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seed}}{The seed (or seeds) that one wishes to retrieve.}

\item{\code{.filtertext}}{The text to add to the filter list (mostly used internally).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STS2RunHistory-filter_character"></a>}}
\if{latex}{\out{\hypertarget{method-STS2RunHistory-filter_character}{}}}
\subsection{Method \code{filter_character()}}{
Retrieve runs containing character/s across the run history.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STS2RunHistory$filter_character(
  char,
  onlyowner = FALSE,
  .filtertext = "filtered by character"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{char}}{The character/s to retrieve data for.}

\item{\code{onlyowner}}{If TRUE, only retrieve runs where the owner was the character specified.}

\item{\code{.filtertext}}{The text to add to the filter list (mostly used internally).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STS2RunHistory-filter_outcome"></a>}}
\if{latex}{\out{\hypertarget{method-STS2RunHistory-filter_outcome}{}}}
\subsection{Method \code{filter_outcome()}}{
Retrieve runs with desired outcome/s across the run history.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STS2RunHistory$filter_outcome(outcome, .filtertext = "filtered by outcome")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{outcome}}{The outcome/s to retrieve data for.}

\item{\code{.filtertext}}{The text to add to the filter list (mostly used internally).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STS2RunHistory-filter_ascension"></a>}}
\if{latex}{\out{\hypertarget{method-STS2RunHistory-filter_ascension}{}}}
\subsection{Method \code{filter_ascension()}}{
Retrieve runs with desired ascensions across the run history.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STS2RunHistory$filter_ascension(
  ascension = 0,
  .filtertext = "filtered by ascension"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ascension}}{The ascensions to retrieve data for.}

\item{\code{.filtertext}}{The text to add to the filter list (mostly used internally).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STS2RunHistory-filter_playercount"></a>}}
\if{latex}{\out{\hypertarget{method-STS2RunHistory-filter_playercount}{}}}
\subsection{Method \code{filter_playercount()}}{
Retrieve runs with desired player count across the run history.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STS2RunHistory$filter_playercount(
  players = 1,
  .filtertext = "filtered by player count"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{players}}{The player count/s to retrieve data for.}

\item{\code{.filtertext}}{The text to add to the filter list (mostly used internally).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STS2RunHistory-filter_version"></a>}}
\if{latex}{\out{\hypertarget{method-STS2RunHistory-filter_version}{}}}
\subsection{Method \code{filter_version()}}{
Retrieve runs with desired patch version across the run history.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STS2RunHistory$filter_version(
  cond = "==",
  patch,
  .filtertext = "filtered by version"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cond}}{A condition (e.g. "=" or ">").}

\item{\code{patch}}{The version to compare runs against.}

\item{\code{.filtertext}}{The text to add to the filter list (mostly used internally).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STS2RunHistory-filter_floorcount"></a>}}
\if{latex}{\out{\hypertarget{method-STS2RunHistory-filter_floorcount}{}}}
\subsection{Method \code{filter_floorcount()}}{
Retrieve runs with desired floor count across the run history.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STS2RunHistory$filter_floorcount(
  floors,
  .filtertext = "filtered by number of floors"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{floors}}{The floor count/s to retrieve data for.}

\item{\code{.filtertext}}{The text to add to the filter list (mostly used internally).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STS2RunHistory-generate_summary"></a>}}
\if{latex}{\out{\hypertarget{method-STS2RunHistory-generate_summary}{}}}
\subsection{Method \code{generate_summary()}}{
Generate a summary dataframe for the run history.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STS2RunHistory$generate_summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STS2RunHistory-clone"></a>}}
\if{latex}{\out{\hypertarget{method-STS2RunHistory-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STS2RunHistory$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
